// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  TEACHER: 'TEACHER',
  STUDENT: 'STUDENT',
  CLASS_MANAGER: 'CLASS_MANAGER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const SubmissionType = {
  ONLINE_TEXT: 'ONLINE_TEXT',
  FILE_UPLOAD: 'FILE_UPLOAD',
  BOTH: 'BOTH',
  NONE: 'NONE'
} as const

export type SubmissionType = (typeof SubmissionType)[keyof typeof SubmissionType]


export const AttendanceStatus = {
  PRESENT: 'PRESENT',
  ABSENT: 'ABSENT',
  LATE: 'LATE',
  EXCUSED: 'EXCUSED'
} as const

export type AttendanceStatus = (typeof AttendanceStatus)[keyof typeof AttendanceStatus]


export const ReportType = {
  STUDENT_ATTENDANCE: 'STUDENT_ATTENDANCE',
  TEACHER_ATTENDANCE: 'TEACHER_ATTENDANCE',
  COURSE_ENROLLMENT: 'COURSE_ENROLLMENT',
  ACADEMIC_PERFORMANCE: 'ACADEMIC_PERFORMANCE',
  DEPARTMENT_STATS: 'DEPARTMENT_STATS',
  CLASS_SCHEDULE: 'CLASS_SCHEDULE',
  CUSTOM: 'CUSTOM'
} as const

export type ReportType = (typeof ReportType)[keyof typeof ReportType]